services/embeddings/requirements.txt
google-cloud-storage
torch